requirements.txt
cffi==1.9.1
cryptography==1.7.1
dmidecode==0.8.1
enum34==1.1.6
idna==2.1
ipaddress==1.0.17
pkg-resources==0.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
PyKCS11==1.3.3
pyOpenSSL==16.2.0
six==1.10.0
urllib3==1.19.1
uuid==1.30
